pyproject.toml
[project]
name = "attendee-realtime-transcription-pipecat"
version = "0.1.0"
description = "Real-time voice agent using Pipecat framework with Attendee integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.68.0",
    "uvicorn>=0.15.0",
    "websockets>=10.0",
    "python-multipart>=0.0.5",
    "python-dotenv>=0.19.0",
    "httpx>=0.23.0",
    "pipecat-ai[daily,elevenlabs,openai,silero,google,runner]>=0.0.79",
    "pyngrok>=7.0.0",
    "jinja2>=3.0.0",
]
